[refactor] #79 ContentDetail.Content -> BaseContentDetail로 분리
